Run adapters in a separate job on GitHub Actions
Closes #25892
