+++ .github/actions/build-keycloak/action.yml
@@ -32,36 +32,12 @@ runs:
       name: Frontend Plugin Cache
       uses: ./.github/actions/frontend-plugin-cache
 
-    # Remove once https://github.com/keycloak/keycloak/issues/19299 is solved
-    ########################################################################################################
-    - id: check-adapter-changes
-      if: github.event_name == 'pull_request'
-      name: Check changes for WildFly adapters
-      shell: bash
-      # If there are no changes for WildFly adapters, we use adapters built in the latest nightly build
-      run: |
-        WF_ADAPTERS_REGEX="^adapters/oidc/wildfly|^adapters/saml/wildfly"
-        
-        git fetch origin --tags --force
-        
-        echo "GIT_WF_ADAPTERS_DIFF=$(git diff origin/main --name-only | egrep -ic -e "$WF_ADAPTERS_REGEX")" >> $GITHUB_ENV
-        echo "NIGHTLY_DIFF=$(git diff nightly --name-only | egrep -ic -e "$WF_ADAPTERS_REGEX")" >> $GITHUB_ENV
-
-    - id: set-maven-profile
-      if: ${{ github.event_name != 'pull_request' || env.GIT_WF_ADAPTERS_DIFF != 0 || env.NIGHTLY_DIFF != 0}}
-      name: Set profile for building distribution
-      shell: bash
-      run: |
-        echo "MVN_PROFILES=-Pdistribution" >> $GITHUB_ENV
-        echo "WildFly adapters will be built in our codebase"
-    ########################################################################################################
-
     - id: build-keycloak
       name: Build Keycloak
       shell: bash
       # By using "dependency:resolve", it will download all dependencies used in later stages for running the tests
       run: |
-        ./mvnw install dependency:resolve -V -e -DskipTests -DskipExamples ${{ env.MVN_PROFILES}}
+        ./mvnw install dependency:resolve -V -e -DskipTests -DskipExamples
 
     - id: compress-keycloak-maven-repository
       name: Compress Keycloak Maven artifacts
