+++ testsuite/integration-arquillian/tests/base/pom.xml
@@ -513,7 +513,6 @@
         <profile>
             <id>app-server-wildfly</id>
             <activation>
-                <activeByDefault>true</activeByDefault>
                 <property>
                     <name>app.server</name>
                     <value>wildfly</value>
