+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/adapter/authz/example/PermissiveModeAdapterTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Red Hat, Inc. and/or its affiliates
+ * Copyright 2016 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,16 +14,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.testsuite.authz.adapter.example;
+package org.keycloak.testsuite.adapter.authz.example;
+
+import java.io.File;
+import java.io.IOException;
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Test;
-
-import java.io.File;
-import java.io.IOException;
+import org.keycloak.testsuite.arquillian.AppServerTestEnricher;
 import org.keycloak.testsuite.arquillian.annotation.AppServerContainer;
 import org.keycloak.testsuite.utils.arquillian.ContainerConstants;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
+
 
 /**
  * @author <a href="mailto:psilva@redhat.com">Pedro Igor</a>
@@ -35,24 +42,30 @@ import org.keycloak.testsuite.utils.arquillian.ContainerConstants;
 @AppServerContainer(ContainerConstants.APP_SERVER_EAP71)
 @AppServerContainer(ContainerConstants.APP_SERVER_TOMCAT8)
 @AppServerContainer(ContainerConstants.APP_SERVER_TOMCAT9)
-public class ServletAuthzLazyLoadPathsAdapterTest extends AbstractServletAuthzAdapterTest {
+public class PermissiveModeAdapterTest extends AbstractBaseServletAuthzAdapterTest {
 
     @Deployment(name = RESOURCE_SERVER_ID, managed = false)
     public static WebArchive deployment() throws IOException {
         return exampleDeployment(RESOURCE_SERVER_ID)
-                .addAsWebInfResource(new File(TEST_APPS_HOME_DIR + "/servlet-authz-app/keycloak-lazy-load-authz-service.json"), "keycloak.json");
+                .addAsWebInfResource(new File(TEST_APPS_HOME_DIR + "/servlet-authz-app/servlet-authz-realm.json"), "keycloak-permissive-authz-service.json");
     }
 
     @Test
-    public void testPathPEPDisabled() {
+    public void testCanAccessWhenPermissive() throws Exception {
         performTests(() -> {
-            login("alice", "alice");
-            assertWasNotDenied();
+            login("jdoe", "jdoe");
+            driver.navigate().to(getResourceServerUrl() + "/enforcing/resource");
 
-            navigateTo();
-            getLink("PEP Disabled").click();
+            if (AppServerTestEnricher.isEAP6AppServer() || AppServerTestEnricher.isTomcatAppServer()) {
+                assertThat(driver.getPageSource(), containsString("HTTP Status 404"));
+            } else {
+                assertThat(driver.getTitle(), is(equalTo("Error")));
+                assertThat(driver.getPageSource(), containsString("Not Found"));
+            }
 
-            hasText("Policy enforcement is disabled. Access granted: true");
+            driver.navigate().to(getResourceServerUrl() + "/protected/admin");
+            assertWasDenied();
         });
     }
+
 }
