+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/adapter/authz/example/ServletAuthzCacheDisabledAdapterTest.java
@@ -14,15 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.testsuite.authz.adapter.example;
+package org.keycloak.testsuite.adapter.authz.example;
 
-import java.io.File;
-import java.io.IOException;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Test;
 import org.keycloak.representations.idm.authorization.ResourcePermissionRepresentation;
 import org.keycloak.representations.idm.authorization.ResourceRepresentation;
+
+import java.io.File;
+import java.io.IOException;
 import org.keycloak.testsuite.arquillian.annotation.AppServerContainer;
 import org.keycloak.testsuite.utils.arquillian.ContainerConstants;
 
@@ -36,16 +37,16 @@ import org.keycloak.testsuite.utils.arquillian.ContainerConstants;
 @AppServerContainer(ContainerConstants.APP_SERVER_EAP71)
 @AppServerContainer(ContainerConstants.APP_SERVER_TOMCAT8)
 @AppServerContainer(ContainerConstants.APP_SERVER_TOMCAT9)
-public class ServletAuthzCacheLifespanAdapterTest extends AbstractServletAuthzAdapterTest {
+public class ServletAuthzCacheDisabledAdapterTest extends AbstractServletAuthzAdapterTest {
 
     @Deployment(name = RESOURCE_SERVER_ID, managed = false)
     public static WebArchive deployment() throws IOException {
         return exampleDeployment(RESOURCE_SERVER_ID)
-                .addAsWebInfResource(new File(TEST_APPS_HOME_DIR + "/servlet-authz-app/keycloak-cache-lifespan-authz-service.json"), "keycloak.json");
+                .addAsWebInfResource(new File(TEST_APPS_HOME_DIR + "/servlet-authz-app/keycloak-cache-disabled-authz-service.json"), "keycloak.json");
     }
 
     @Test
-    public void testCreateNewResourceWaitExpiration() {
+    public void testCreateNewResource() {
         performTests(() -> {
             login("alice", "alice");
             assertWasNotDenied();
@@ -66,17 +67,7 @@ public class ServletAuthzCacheLifespanAdapterTest extends AbstractServletAuthzAd
             permission.addResource(resource.getName());
             permission.addPolicy("Deny Policy");
 
-            getAuthorizationResource().permissions().resource().create(permission).readEntity(ResourcePermissionRepresentation.class);
-
-            login("alice", "alice");
-            assertWasNotDenied();
-
-            this.driver.navigate().to(getResourceServerUrl() + "/new-resource");
-            assertWasNotDenied();
-
-            //Thread.sleep(5000);
-            setTimeOffset(30);
-            setTimeOffsetOfAdapter(30);
+            permission = getAuthorizationResource().permissions().resource().create(permission).readEntity(ResourcePermissionRepresentation.class);
 
             login("alice", "alice");
             assertWasNotDenied();
@@ -84,12 +75,18 @@ public class ServletAuthzCacheLifespanAdapterTest extends AbstractServletAuthzAd
             this.driver.navigate().to(getResourceServerUrl() + "/new-resource");
             assertWasDenied();
 
-            resetTimeOffset();
-            setTimeOffsetOfAdapter(0);
-        });
-    }
+            permission = getAuthorizationResource().permissions().resource().findById(permission.getId()).toRepresentation();
 
-    public void setTimeOffsetOfAdapter(int offset) {
-        this.driver.navigate().to(getResourceServerUrl() + "/timeOffset.jsp?offset=" + String.valueOf(offset));
+            permission.removePolicy("Deny Policy");
+            permission.addPolicy("Any User Policy");
+
+            getAuthorizationResource().permissions().resource().findById(permission.getId()).update(permission);
+
+            login("alice", "alice");
+            assertWasNotDenied();
+
+            this.driver.navigate().to(getResourceServerUrl() + "/new-resource");
+            assertWasNotDenied();
+        });
     }
 }
