+++ testsuite/integration-arquillian/tests/base/src/test/java/org/keycloak/testsuite/adapter/authz/example/ServletAuthzCacheLifespanAdapterTest.java
@@ -14,16 +14,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.testsuite.authz.adapter.example;
+package org.keycloak.testsuite.adapter.authz.example;
 
+import java.io.File;
+import java.io.IOException;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
 import org.junit.Test;
 import org.keycloak.representations.idm.authorization.ResourcePermissionRepresentation;
 import org.keycloak.representations.idm.authorization.ResourceRepresentation;
-
-import java.io.File;
-import java.io.IOException;
 import org.keycloak.testsuite.arquillian.annotation.AppServerContainer;
 import org.keycloak.testsuite.utils.arquillian.ContainerConstants;
 
@@ -37,16 +36,16 @@ import org.keycloak.testsuite.utils.arquillian.ContainerConstants;
 @AppServerContainer(ContainerConstants.APP_SERVER_EAP71)
 @AppServerContainer(ContainerConstants.APP_SERVER_TOMCAT8)
 @AppServerContainer(ContainerConstants.APP_SERVER_TOMCAT9)
-public class ServletAuthzCacheDisabledAdapterTest extends AbstractServletAuthzAdapterTest {
+public class ServletAuthzCacheLifespanAdapterTest extends AbstractServletAuthzAdapterTest {
 
     @Deployment(name = RESOURCE_SERVER_ID, managed = false)
     public static WebArchive deployment() throws IOException {
         return exampleDeployment(RESOURCE_SERVER_ID)
-                .addAsWebInfResource(new File(TEST_APPS_HOME_DIR + "/servlet-authz-app/keycloak-cache-disabled-authz-service.json"), "keycloak.json");
+                .addAsWebInfResource(new File(TEST_APPS_HOME_DIR + "/servlet-authz-app/keycloak-cache-lifespan-authz-service.json"), "keycloak.json");
     }
 
     @Test
-    public void testCreateNewResource() {
+    public void testCreateNewResourceWaitExpiration() {
         performTests(() -> {
             login("alice", "alice");
             assertWasNotDenied();
@@ -67,7 +66,17 @@ public class ServletAuthzCacheDisabledAdapterTest extends AbstractServletAuthzAd
             permission.addResource(resource.getName());
             permission.addPolicy("Deny Policy");
 
-            permission = getAuthorizationResource().permissions().resource().create(permission).readEntity(ResourcePermissionRepresentation.class);
+            getAuthorizationResource().permissions().resource().create(permission).readEntity(ResourcePermissionRepresentation.class);
+
+            login("alice", "alice");
+            assertWasNotDenied();
+
+            this.driver.navigate().to(getResourceServerUrl() + "/new-resource");
+            assertWasNotDenied();
+
+            //Thread.sleep(5000);
+            setTimeOffset(30);
+            setTimeOffsetOfAdapter(30);
 
             login("alice", "alice");
             assertWasNotDenied();
@@ -75,18 +84,12 @@ public class ServletAuthzCacheDisabledAdapterTest extends AbstractServletAuthzAd
             this.driver.navigate().to(getResourceServerUrl() + "/new-resource");
             assertWasDenied();
 
-            permission = getAuthorizationResource().permissions().resource().findById(permission.getId()).toRepresentation();
-
-            permission.removePolicy("Deny Policy");
-            permission.addPolicy("Any User Policy");
-
-            getAuthorizationResource().permissions().resource().findById(permission.getId()).update(permission);
-
-            login("alice", "alice");
-            assertWasNotDenied();
-
-            this.driver.navigate().to(getResourceServerUrl() + "/new-resource");
-            assertWasNotDenied();
+            resetTimeOffset();
+            setTimeOffsetOfAdapter(0);
         });
+    }
+
+    public void setTimeOffsetOfAdapter(int offset) {
+        this.driver.navigate().to(getResourceServerUrl() + "/timeOffset.jsp?offset=" + String.valueOf(offset));
     }
 }
