+++ testsuite/integration-arquillian/tests/base/testsuites/base-suite
@@ -1,6 +1,6 @@
 account,4
 actions,1
-adapter,2
+adapter,IGNORED
 admin,1
 authz,3
 broker,2
